Google GPS Application for Android 6
This application uses the Google Maps API for Android. It features a
satellite map, with the ability to add and remove markers through the
touch screen. These markers display the coordinates and address (when
available). There is a My Location button which displays the user's
current location with a marker automatically added. On the bottom of the
application, there are two buttons, one to remove the marker which the
user has currently selected, and the other removes all markers on the
map. There also are text fields showing the user's current moving speed,
altitude, and the accuracy of the GPS.